[components] Merge https://github.com/mozilla-mobile/android-components/pull/8049
8049: For https://github.com/mozilla-mobile/android-components/issues/6099 - Convert config to yaml r=pocmo a=NotWoods

In Gradle 6, buildSrc classes cannot be used inside settings.gradle. Inlining the `Config` class gets around this problem for us. The values inside the class have been moved to a YAML file so that it can be used by Taskcluster to check if data has changed, and used by the changelog to be a simple version overview.

We could also use a .properties file for this.

Co-authored-by: Tiger Oakes <toakesmozilla.com>

UltraBlame original commit: 3af549cc04ced7352233b5477ff69c5600c613ba