[components] Closes https://github.com/mozilla-mobile/android-components/issues/8497: Make RecentlyClosedMiddleware listen to UndoAction.ClearRecoverableTabs.
With that closed tabs will not immediately show up in "closed tab" and instead we will
wait for the "undo" timeout and only if the user didn't "undo" then the tab will move
to the list of closed tabs.

Previously selecting "undo" caused the tab to exist in the tabs tray as well as in
the list of closed tabs.

UltraBlame original commit: 0e86a234bf057829ab7cc203be6abbb577aeaaf0