[components] For https://github.com/mozilla-mobile/android-components/issues/9189 - Refactor service-pocket to support recommended articles.
Since the previous version of this component was not in use anymore it was used
as a starting point for the newly needed functionality - easily offering Pocket
recommended articles to interested clients.

The new version offers a simple API for downloading, caching and refreshing in
the background the list of Pocket articles that is offered to clients.

UltraBlame original commit: 2575e6188230739fd63d4ab0f574cc85d71efa9b