[components] For https://github.com/mozilla-mobile/android-components/issues/10749 - New NonBlockingHttpIconLoader
This will download and persist the network resources in the background without
blocking actually returning a response.
This means the first request for a network resource will result in a generated
icon but future requests of the same resources should be able to load the
locally persisted icon.

UltraBlame original commit: 9c613f10c2b764bbbcf46ba9dde41cf9fd572157