[components] For https://github.com/mozilla-mobile/android-components/issues/7477 - Execute WebNotification click based on incoming Intent
Gecko's WebNotification will be persisted in the native Android notification
that is posted like before to target a specific activity in the app.
When the user clicks the notification the target activity can delegate a new
WebNotificationIntentProcessor component to actually process notification's
contentIntent and trigger `click` on the WebNotification.

UltraBlame original commit: 96640fcbd6308244bfd0dfa3750064587dcf656a