[components] Bug 1782194 - Enforce a single Glean major version across all dependencies
Previously we had a check in the Glean Gradle plugin that tried to
enfore a _single_ Glean version across the whole code base.
This is accompanied by a change to the Glean Gradle plugin.

That check wasn't working inside Android Components to begin with,
because it runs on individual components and there's no component which depends on both GeckoView (which provides glean-native)
and an app-services component that also depends on Glean.
It therefore didn't really catch things most of the time.

But when it did catch things (and I have yet to understand exactly when that would be the case)
it really got more in our way than necessary:
To ensure we didn't run into version issues we had to always upgrade Glean in appservices
and get a release before we land upgrades of GeckoView, Glean _and_ appservices in a-c.
That's needlessly restrictive.

What we really want:
The version exported (as a capability) from GeckoView to be used in the final build.
And no other component using a Glean major version lower or higher than that.
Glean itself follows SemVer and so major versions should be compatible.
Enforcing the version exported by GeckoView is therefore enough to get a working build.

Unfortunately in Gradle we can't really ask what version GeckoView provides,
but luckily for use we already have `Versions.mozilla_glean`, which is auto-updated.
We should just not manually modify that value ever.

Note: for major version upgrades we _still_ need to get the Glean update
into appservices first, get that released and then update in A-C.
Hopefully this happens less often now.

UltraBlame original commit: 1430101df805dd0f45dda87d086d8e66c1f8ded3