[components] For https://github.com/mozilla-mobile/android-components/issues/12882: Allow the QR functionality to be queried an resumed.
Clients can query the QRFragment for isScanInProgress and decide on whether to
resume scanning by calling `scan` in a new `QRFeature` instance.
This may be needed if the process is restarted as a followup to the user
updating system permissions for the app which would recreate the component
owning the QRFeature.

UltraBlame original commit: daa23c19d2ebb022e9368bca04c3c0124ca92a8e