Bug 1795903 - Introduce StorageMaintenanceWorker that runs periodically to prune places storage.
WorkManager's Worker is created for periodically running storage maintenance operation on places database.
DB size limit is exposed from A-S that indicates the maximum DB size to aim for, in bytes. If the database
exceeds this size, a small number of visits will be pruned. This operation is done periodically by a periodic
work request that, optimally, will run once a day as long as the worker constraints are met.

UltraBlame original commit: c33432679b5a90b5fe8de25bbc3f4252917c93d1