[focus] Correctly track visibility state of MainActivity. (https://github.com/mozilla-mobile/focus-android/issues/774)
When clicking the notification to erase the browsing history our BrowsingNotificationService
needs to decide:

* Is the app in the background? Then close it so that we start fresh the next time it is started.
* Is it in the foreground? Then notify it and play the erase animation.

The previous version tracked forground/background state by looking at onResume() and onPause().
This is not correct in all cases: An activity can be visible but not in the foreground (e.g.
another translucent screen is on top). In this case we still want to play the animation and
not (visibly) close the app.

With this patch we now track the state from onStart() and onStop(). Those callbacks are exactly
for this: They are called once an activity becomes visible to the user.

Additionally with this patch we will track the start of a browsing session as a "foreground"
event. This is necessary because since the service could have been removed since the last
foreground event (it's only "ongoing" *during* a session). However we are *always* in the
foreground when starting a browsing session.

UltraBlame original commit: ee7a792a08c33d1586849c2428a1a4bce541de98