[focus] Fixes https://github.com/mozilla-mobile/focus-android/issues/2194 - Add support for new GeckoView pin-to-screen API
GeckoView now has an API to let the consumer know when the view should
be pinned to its current location on the screen. We use this new API in
NestedGeckoView and prevent scrolling of the view when the API returns
true.

UltraBlame original commit: 3b229e21959c2a92321a459bdf31f5352f1b0136