Bug 1810150 - stop scheduling pull-request task graphs when there's already a corresponding push graph
When a PR is opened from a branch in the main repo, we used to create a
task graph for both the push event and the pull-request open or
synchronize event, associated with the same commit, and thus essentially
doing the same work twice.

With this change, we skip the pull-request graph, and rely on the push
graph to run the required checks.  This requires corresponding changes
to the mergify config to consider either of complete-pr or complete-push
as good enough to merge a PR.

UltraBlame original commit: 061418f1fc05d2daf0adfcd4474155fb2548d70c