[fenix] For https://github.com/mozilla-mobile/fenix/issues/6126 - moved the sessionObserver init to the onCreate
This helps because we will always need the observer to be initiated, not only when the `openToBrowser` method gets called. Example: Opening a tab from the tab tray had it's own method for opening the browser, causing this to not be called.

UltraBlame original commit: e402d91d398723b498c7008ab24c446e480a3adb