[fenix] For https://github.com/mozilla-mobile/fenix/issues/13959: change StrictModeManager to class from object.
I originally tried to create this PR leaving this as an object to keep
the change simple but it wasn't worth it - once the object started to
keep state, we'd need to manually reset the state between runs. Also,
the tests were already getting hacky with static mocking so it was
easier to address some of those issues this way too.

UltraBlame original commit: 97b8b44212874b68d3df37e3f5921bd387d24ef4