[fenix] For https://github.com/mozilla-mobile/fenix/issues/18122 - Update browser layout for find in page bar with bottom toolbar
With a dynamic bottom toolbar the  browser extends all the way to the bottom of
the screen and so when the find in page bar is shown it will be so above the
bottom part of the browser which may contain exactly the searched for items.

To fix the browser must effectively be placed at the top of the newly shown
find in page bar.

UltraBlame original commit: aafe351c6560de3d9e63f7433a44c5402e51b48c