[fenix] Close https://github.com/mozilla-mobile/fenix/issues/18931: Implement add to collections in interactor
We moved the collection dialog code out from the old fragment, because it
had nothing to do with tabs tray, and into the collections package to be
re-usable in other parts of the app.

In addition, we also make use of it in the new tabs tray's
NavigationInteractor.

UltraBlame original commit: bfe264be9558cc06360964e3c0016ff3ff8bd896