[fenix] For https://github.com/mozilla-mobile/fenix/issues/18591 - Update tests
Some tests trying to "Add to top sites" may fail because of the sticky navbar
now covering that menu item -> scroll the menu up a little.

Some tests that previously were trying to scroll to the bottom of the menu to
interact with navbar options now don't have to. Them trying to interact with
navbar option with the menu opened as collapsed is a good test for the navbar
stickyness.

UltraBlame original commit: bc845d0669f613a30f9118027efc5b5a74737f71