[fenix] For https://github.com/mozilla-mobile/fenix/issues/19135 - Set different offsets for expanded trays
As in the previous version of tabs tray when tray is expanded while in portrait
it should have a 40dp offset and when the tray is expanded while in landscape
it should have a 0dp offset - effectively making it fullscreen.

These values were set in code, always overriding the xml set value so I
removed that xml declaration.

UltraBlame original commit: ef350f566ef5d64dee937d00fdc1368edf960a16