[fenix] For https://github.com/mozilla-mobile/fenix/issues/21183: remove view hierarchy depth check from excessive resource test.
This doesn't seem to be a high value test: increasing the view hierarchy
depth will only result in a performance problem on low end devices
if there is enough content on the new layer to cause the traversal to
take longer. It's more likely to result in a hard-to-workaround false
positive so we can remove it, like component init count.

UltraBlame original commit: f16a8f71a7eb0710684aa9196dee7acd60b4125a