[fenix] Convert recently closed tabs code to use light-weight TabState
To lighten-up our memory usage and startup performance, all of the RecentlyClosed
machinery was converted to use a light-weight TabState - specifically, it's missing
EngineSessionState, which is expensive to obtain during startup, and potentially
very costly to keep in-memory.

When we actually need EngineSessionState (at the point of restoration of a tab), we
read and rehydrate it using provided storage implementation.

UltraBlame original commit: 9c54968e109500b5cf4c71a72063e094900a7f39