[fenix] For https://github.com/mozilla-mobile/fenix/issues/11753 - Update compose to 1.1.1 and Kotlin to 1.6.10
The needed changes are around supporting exhaustive whens for sealed classes
as a new kotlin feature - https://kotlinlang.org/docs/whatsnew1530.html#exhaustive-when-statements-for-sealed-and-boolean-subjects

androidx_activity_compose was removed as a dependency since it isn't used.

Used 1.6.10 for Kotlin although 1.6.20 is available to prevent any issues with
Compose 1.1.1 reported as an error at compile time:
"e: This version (1.1.1) of the Compose Compiler requires Kotlin version 1.6.10
but you appear to be using Kotlin version 1.6.20 which is not known to be
compatible. Please fix your configuration (or
`suppressKotlinVersionCompatibilityCheck` but don't say I didn't warn you!)."

UltraBlame original commit: 6d41caac9565ab081bcf88a556b468e2106871ac