[fenix] For https://github.com/mozilla-mobile/fenix/issues/27627 - Check the TCP CFR status lazily
Use lazyFeatureFlagPreference as a quick small way to avoid a race between
initializing the value and the Nimbus initialization based on which the value
should be calculated.
This is the same flow that the other MR experiments use.

UltraBlame original commit: e602ab3fccb8f059516e14ccf56443c3d5bac851