[fenix] For https://github.com/mozilla-mobile/fenix/issues/25816: Support changing autocomplete providers.
Everywhere the toolbar which needs to show autocomplete suggestions is used we
show a `ToolbarView`. So instead of having this configurable in 2 or more
places as it happened before the autocomplete functionality is configured only
from the `ToolbarView` class.
It will contain a `ToolbarAutocompleteFeature` that will be immediately updated
with the appropriate autocomplete providers or remove all such providers to
immediately update the current autocomplete or remove it entirely depending on
the new search engine selected.

UltraBlame original commit: c994b6e97b198e11440f6e80f01ae1a32c16036d