Bug 1837526 - Call StartupTimeline in the Compose Top Sites
Because we are migrating Top Sites to Compose, we want to ensure we
are still recording when top sites are loaded, therefore the call
for StartupTimeline is now called after creating the top sites
compose elements. Calling it here will ensure it will be called
even after we replace the home page recycler with a Compose view.

UltraBlame original commit: 13b7572e9e2a028690db486f1ea6ff68dac5609c