Bug 1875294 - Record breadbcrumbs before crashing with UnsatsisfiedLinkError
Report breadcrumbs to Sentry before crashing with
`UnsatsisfiedLinkError`:

  - The files inside the app directory, to detect if the library is
    present or not.
  - The files inside the APK, to detect if a file inside the APK didn't
    get installed correctly.
  - The name of the installer package, so that we can know if the APK
    was installed from a 3rd-party source.
  - The path to the APK.  I don't really see this being useful, but it's
    needed to determine all the other info, so we might as well report
    it.

The idea here is that hopefully one of these will detect that something
was off with their install.  For example, maybe they installed an APK
for the wrong arch or maybe `libjnidispatch` was not extracted
correctly.  If not, at least we can rule these possibilities out as the
root cause.

UltraBlame original commit: 4097b932b08aafcab3c7614c4bb0ef6fd08b4587