Bug 1809798 - Loads the right wallpaper when changing orientation
Because of the android:configChanges from Manifest that disable the
automate Activity recreation, the responsibility of handling the
configuration changes is transferred to Activity. When configurations
changes, onConfigurationChanged() is called.

So now when this function called, the correct orientation is sent
forward as parameter of applyWallpaper.

UltraBlame original commit: 4751c85f681d0239b7fdc0871661c968ce96eb87