Bug 1880483 — Add inverse triggers to messaging component
This is a breaking change.

This changes the `trigger` key to `trigger-if-all` to document that all
triggers in this list must evaluate to true.

It also adds `exclude-if-any`, a list of triggers which, if any are true
will exclude the message.

This will remove the need for triggers like `I_AM_NOT_BROWSER_DEFAULT`.

UltraBlame original commit: dc0eb1595e1aed746b281c5d9c7726d40a3a37e4