Bug 1882333 - [devtools] Update the editor toggle button size when the editior dimensions change r=devtools-reviewers,nchevobbe
This patch updates the size of the button (above the editor gutter) which toogles the sidebar for codemirror 6.
Note this was broken with the current codemirror as wel so fixed both.

Also adding the ability to attach an updateListener to the editor we can listen for editor changes.

Differential Revision: https://phabricator.services.mozilla.com/D202398

UltraBlame original commit: ada2ec6c4fa202d1456805375ed796ac2f21f3b7