Bug 1882789 - Register the LoadedScript in the memory reporter. r=arai
This change adds the `LoadedScript` in the memory reporter.

This change is needed as we are looking into using the `LoadedScript` as a value
of the `SharedSubResourceCache`. The upcoming cache requires the Value to be
registered in the memory reporter, providing a `SizeOfInlcudingThis` function.

Differential Revision: https://phabricator.services.mozilla.com/D203132

UltraBlame original commit: 0c2718416644ed2456d14e38f10496ea6685b015