Bug 1885972 - Remove 'persistent-usage-record' MediaKeySessionType r=media-playback-reviewers,padenot
This patch removes the `cdm::SessionType::kPersistentUsageRecord`.

It was introduced in June 2020 to comply with the EME spec #462 [1],
although Mozilla's position [2] was ambiguous at the time. Subsequently,
in January 2021, the `persistent-usage-record` was removed from
`MediaKeySessionType` in #482 [3] for issue #480 [4]. However, the
addition of `kPersistentUsageRecord` was not removed accordingly, until
this patch.

[1] https://github.com/w3c/encrypted-media/pull/462
[2] https://github.com/mozilla/standards-positions/issues/407
[3] https://github.com/w3c/encrypted-media/pull/482
[4] https://github.com/w3c/encrypted-media/issues/480

Differential Revision: https://phabricator.services.mozilla.com/D204965

UltraBlame original commit: a72682a2fa9b5c753d697dfbda30da954c0c5201