Bug 1886654 - Initialize content PDMs when RDD process is not enabled. r=media-playback-reviewers,padenot
On platforms where the RDD process is not enabled (Android, Solaris,
NetBSD, DragonflyBSD, iOS), video decoding is rejected because the
content PDMs are not enabled when media.allow-audio-non-utility is
false (which it is), except on Android because the
media.allow-audio-non-utility check is ifdef'ed out currently.

Instead of relying on ifdefs, rely on whether the RDD process is
enabled.

Differential Revision: https://phabricator.services.mozilla.com/D205282

UltraBlame original commit: 131cc12d8e2eceac8d44d4a02bb56f71764946d5