Bug 1870617 - Map Windows error 0x1 ERROR_INVALID_FUNCTION to NS_ERROR_FILE_DEVICE_FAILURE r=jstutte,xpcom-reviewers
`ERROR_INVALID_FUNCTION` appears to be a very generic "this device does not support this function".
Given that we deal only with file operations and errors here, this pretty much translates to "this device is not really a working file IO device for us" which makes it appear to be entirely broken for us.

Differential Revision: https://phabricator.services.mozilla.com/D196723

UltraBlame original commit: 33bfeab02910fb736732b4ff3009d4daba7b09a5