Bug 1867058 - Part 2: Make both StaticRange and nsRange to have a way to cross the trees r=jjaschke,smaug,dom-core
For StaticRange, simply allows node to cross the trees.

For nsRange, when nodes are in different trees, in addition
to collapse them one point, a StaticRange will be created and the
original nodes will be stored to this StaticRange. This approach
allows us to have different behaviors of nsRange when the expectation
is different. For most of the cases where nsRange is being used (ie.
Editor), we keep the old behavior such that nodes always in the same
tree. And we can use the new behavior for things like selection.

Differential Revision: https://phabricator.services.mozilla.com/D195302

UltraBlame original commit: 713c66e124c74d0d9c1dda6821bd2984d6cb4040