Bug 1884360 part 3 - Add infrastructure for trampoline natives. r=iain
This patch adds code for using the JitEntry mechanism for native builtins.

The TrampolineNative JSJitInfo mechanism is very similar to what we do for
InlinableNative. For each trampoline native, a JIT trampoline is generated and
we use this as the function's JitEntry when creating the native JSFunction.

Differential Revision: https://phabricator.services.mozilla.com/D204625

UltraBlame original commit: 9ef13ed79db6e509934113d1d3a9a324ff7e6d69