Bug 1888706 - Prevent macOS native titlebar from being completely white. r=desktop-theme-reviewers,jules
This is subtle enough to not be an issue with darker wallpapers, but
prevent the contrast issue where the native titlebar can't be
distinguished from our toolbar.

This is the same tint we use on Linux and macOS for the urlbar.

Differential Revision: https://phabricator.services.mozilla.com/D206153

UltraBlame original commit: ed0b8d8ed86b412039a9c63335c834d5c7059851