Bug 1888264 - Increase osx CaR build run time. r=perftest-reviewers,afinder
Build times for CaR OSX (x64) have increased quite a bit lately and is
slowly reaching the current max-run-time of 35000. This patch bumps it
up to 40000 to keep us in the clear for a while. Eventually we will be
able to build these faster if Bug 1858740 pans out.

Differential Revision: https://phabricator.services.mozilla.com/D205922

UltraBlame original commit: a6eb5bc4af4569aebf5617bde2573988578fb8a0