Bug 1884442 - Android preference to offer a translation persistence r=android-reviewers,giorga
This patch fixes an issue where the Gecko pref `browser.translations.automaticallyPopup`
was not persisting in Android. The issue was it was being overwritten back to the
default of `true`. This patch stores the persistence in the Fenix layer and
ensures the right value is kept after user interaction.

Differential Revision: https://phabricator.services.mozilla.com/D206268

UltraBlame original commit: 2540d7ed95363b8c0421edff9932c0d2cfcd3eaa