Bug 1867939, part 3: Implement an algorithm to find text directives in a document. r=hsivonen,dom-core
Main part of this patch is an implementation of the algorithm defined
in the spec [0].
The algorithm takes an array of `TextDirective` (which is the result of
the fragment directive parser implemented in part 2 of this bug)
 and converts it into an array of `nsRange`s.

The algorithm is not optimized for performance yet.
Instead it follows the spec very closely.
Therefore, it may for now only be seen as a baseline to a faster `nsFind` based approach.
The algorithm differs from the spec, using a standard `nsString::Find()`
instead of the mentioned collator-based search, i.e.
performing a case-sensitive and locale-independent search.

[0]: https://wicg.github.io/scroll-to-text-fragment/

Differential Revision: https://phabricator.services.mozilla.com/D195686

UltraBlame original commit: 38631fbd2f2fe46b866d26c8bf865668f8b9d629