Bug 1889091 - Part 6: Remove "scratch" register parameter from emitPushArguments. r=jandem
It was useful to use an explicit parameter when `scratch` was used to hold the
extra stack space. But now that we no longer have to compute the extra stack
space, it's no longer necessary to have an out-param and we can instead use a
local variable similar to how `ToRegister(apply->getTempObject())` is used.

Depends on D206355

Differential Revision: https://phabricator.services.mozilla.com/D206356

UltraBlame original commit: 7a8990de860d4a0922ac30b0611010d7da078439