Bug 1888149 [wpt PR 45373] - screen wake lock: Set "allow" attribute in more Permissions Policy tests, a=testonly
Automatic update from web-platform-tests
screen wake lock: Set "allow" attribute in more Permissions Policy tests (#45373)

Contrary to the old Feature Policy spec, in the Permissions Policy both the
header and the "allow" attribute are checked, and even when the header
access to a given feature it might still fail if the "allow" attribute is
not set accordingly (it defaults to "src").

In the tests being changed here, we explicitly set the attribute to make
sure that the Permissions Policy checks are failing because of the header
values, not because of a lack of an "allow" attribute that would have
otherwise permitted access to the feature.
--

wpt-commits: 25fdec3c95e308d6c844022fedd2fcbb6b2b5703
wpt-pr: 45373

UltraBlame original commit: c4d4fe277ab8821d180e5d58419023d40086980f