Bug 1888880 - Add application icons to the crashreporter r=gsvelto
This adds a window/application icon to the linux and windows GUIs.

The icon was already added on macOS, but it is irrelevant right now as
it's run as a background application.

This also changes out the `lazy_static` crate for `once_cell` throughout
the application.

Differential Revision: https://phabricator.services.mozilla.com/D206392

UltraBlame original commit: c9d66aa89918b85e79cb4af65d29fec2032abd5e