Bug 1870642 - Fix Collection deleted snackbar that overlaps the toolbar r=android-reviewers,007
When a collection was deleted, the snackbar that contains the undobutton was overlapping the search toolbar.

The issue was that there wasn't a view to anchor the snackbar.
The fix was only to change the anchorView attribute from the functionallowUndo from null to snackbarAnchorView.

| Before | After |
| {F6475661} | {F6475744} |

Differential Revision: https://phabricator.services.mozilla.com/D206652

UltraBlame original commit: ad446301fd3335ef22242f963c58504f086503cd