Bug 1889442 - Always update titlebar button positions. r=mac-reviewers,mstange
I introduced this check in bug 1887063 to deal with the titlebar button
position in the page info / places windows (which use drawtitle to use
the native titlebar, and have an empty button region, since they don't
have any button box in the DOM).

However, after that, in the same patch, I wrote the more correct check
(!win.wantsTitleDrawn in [1]), which means this is now unnecessary for
those windows, but forces a titlebar relayout when entering full-screen
mode on macOS Catalina, which seems to be needed.

I discarded all the MOZTitlebarView from introducing the regression,
since the browser window didn't have a titlebar view before my changes
if there is a theme being used (it'd hit [2] with
needTitlebarView=false), so this was really the only remaining bit.

[1]: https://searchfox.org/mozilla-central/rev/c09764753ea40725eb50decad2c51edecbd33308/widget/cocoa/nsCocoaWindow.mm#3020
[2]: https://hg.mozilla.org/mozilla-central/rev/6dd030b8189cbc5e293ad969d0b6887b5dca8793#l13.250

Differential Revision: https://phabricator.services.mozilla.com/D207002

UltraBlame original commit: fd460127e60fed22678801b35d8c5d127da99545