Bug 1858809 - Enforce Worker::PostEventWithOptions invariants. r=bwc,dom-worker-reviewers,smaug
Under debug/fuzzing builds WorkerRunnable will assert in its constructor
if passed a null WorkerPrivate.  While there is no runtime problem under
non-debug builds, it is a reasonable invariant for us to have until we
are able to eliminate to turn all of this into normal runnables.

`RTCRtpScriptTransform::Constructor` currently can get into this situation
where the method is invoked but the Worker is in a closing / terminated
state.  The spec has a "FIXME" to specify what to do in this situation.

This patch implements an early return with an error based on adding a new
`Worker::IsEligibleForMessaging` which is analogous to
`nsIGlobalObject::IsEligibleForMessaging`.  This is potentially
observable since JS getters or proxies passed to the constructor as the
`options` arg will not be serialized in this case, but it is not a
secret to the inherently same-origin Worker whether the Worker is
closing; the same information is available via WebLocks among other
methods.  I chose the simplest path here given the absence of any spec
language or even an issue on the repo that I can see.

Differential Revision: https://phabricator.services.mozilla.com/D205779

UltraBlame original commit: b4e3dcc4aee727bc1db7770ad0f7b8c9191c81c1