Bug 1887332 - Improve handling of WasmToJSJit frames in JSJitProfilingFrameIterator. r=rhunt
The iterator didn't mark itself `done()` when reaching a WasmToJSJit entry frame and
this confused the code in `GetTopProfilingJitFrame`.

We now store the Wasm caller FP separately and mark the iterator as done.
This is less error-prone and is also more consistent with other JS JIT and Wasm
frame iterators.

Differential Revision: https://phabricator.services.mozilla.com/D206991

UltraBlame original commit: 098eed68d43ba5e173fef30649811281dfa5e842