Bug 1889397 [wpt PR 45499] - [line-clamp] Fix crash when explicitly setting `line-clamp: none`, a=testonly
Automatic update from web-platform-tests
[line-clamp] Fix crash when explicitly setting `line-clamp: none`

https://crrev.com/c/5317993 added the `line-clamp` property, whose
valid values are positive integers and the keyword `none`. The
resolved type of this property was `int`, where the keyword `none`
would map to a value of 0.

However, this property definition did not specify an explicit
converter. This meant that when `line-clamp: none` was explicitly set,
the conversion to an `int` would trigger an assertion failure, since
it expects a `CSSPrimitiveValue`, not a `CSSIdentifierValue`.

This patch fixes this by adding the converter
`StyleBuilderConverter::ConvertIntegerOrNone`, which allows the `none`
keyword in addition to integers, and which takes a template argument
for the integer value to use for `none`.

Bug: 40336192
Change-Id: Ide218b2fec75f21908280f475e85b391b7551f7e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5419122
Reviewed-by: Ian Kilpatrick <ikilpatrickchromium.org>
Commit-Queue: Ian Kilpatrick <ikilpatrickchromium.org>
Cr-Commit-Position: refs/heads/main{#1281892}

--

wpt-commits: 2efef6ffa65554e5e6739f27bffde6a4abb3e49d
wpt-pr: 45499

UltraBlame original commit: 5deee3a94b6268d1d92984d1974cedab9d2ea34e