Bug 1890523 - Remove extra space above/below stretchy operators. r=emilio
Gecko always adds .2em space above and below stretchy operators but this
is not part of MathML Core and is causing several WPT failures. The
MATH table does not have any parameter for that so it's likely some
tweak that was added in the early days of MathML that may no longer be
necessary. This patch tentatively removes this feature by placing it
under a disabled-by-default flag.

Differential Revision: https://phabricator.services.mozilla.com/D207014

UltraBlame original commit: 6a7c866d716d4150c8061e2db4715d4c9aa8b7bd