Bug 1890773 - Create a pref to control how letter-spacing is inserted around characters. r=layout-reviewers,emilio
With the pref set to 0 by default, this does not change any existing behavior.

My intention is that a following patch will set it to 2 in Nightly. This will affect
a bunch of reftests, and so test/expectation adjustments will also be needed.

Differential Revision: https://phabricator.services.mozilla.com/D207150

UltraBlame original commit: 1bb52c5a65eb69afcb20947e6625030684b83a11