Bug 1890918 - Make `ContentComparator::LessThan()` check whether the content nodes are different before calling `nsContentUtils::CompareTreePosition` r=emilio
It seems that it's okay to remove `NS_WARN_IF` in
`nsContentUtils::CompareTreePosition`, but in theory, the comparator should
check it first to save the cost of calling the method.  Although it may be
optimized by the compiler.

Differential Revision: https://phabricator.services.mozilla.com/D207242

UltraBlame original commit: e8a43c54c24e62590218e7f638ff70449440d3d6