Bug 1885597 - Remove GitHub synchronization task r=jrmuizel,jnicol,taskgraph-reviewers,releng-reviewers,ahal,bhearsum
This synchronization is now handled in the downstream repositories. In
the case of WebRender by the Servo project and in the case of qcms by a
Gecko-managed GitHub Action. This change removes the github-sync task
from mozilla-central.

Differential Revision: https://phabricator.services.mozilla.com/D204787

UltraBlame original commit: 5f6ed40be0195d8c2188dbb23ba8453f59f71e81