Bug 1884986 - Ensure the control on an empty card on Firefox View would announce the group info. r=fxview-reviewers,Jamie
Since the wrapper of the `<details>` element cannot access its labeling element (included in the slot), we could move the grouping attributes to the `<card-container>` element itself. This would also allow us to remove an additional `<section>` container and reduce the number of nested groups exposed to the Accessibility API, including the case where another `<details>` element is being added for the inner card.

This patch also removes the extra `group` container for an empty card to reduce the verbocity of the announcement and the nesting complexity, while keeping the grouping programmatically indicated (just not duplicated in the case of an empty card where the additional card template is added).

Differential Revision: https://phabricator.services.mozilla.com/D206261

UltraBlame original commit: dbff562c8c7479e73deb5d97b4d5d899c690bb79