Bug 1885361 - Add SkipUnlessMozlint taskgraph optimization strategy. r=taskgraph-reviewers,ahal
Optimization strategy that uses the mozlint configuration files to decide
whether or not to run a job.

In order to support Firefox and Thunderbird, when registering the strategy,
the path to mozlint config files is needed.

Differential Revision: https://phabricator.services.mozilla.com/D204685

UltraBlame original commit: 050d5001389f0c58d3a62dcd1b439c86b74f384a