Bug 1891408 - part 3: Make some callers of `WSScanResult::Point_Deprecated()` use new methods r=m_kato
`Point_Deprecated()` is really error-prone since its result meaning is different
whether the scan direction is backward or forward.  Therefore, if a caller wants
a point in a text node and the direction is only one of them, we can change it
to use `WSScanResult::PointAtReachedContent()` or
`WSScanResult::PointAfterReachedContent()`.

Depends on D207686

Differential Revision: https://phabricator.services.mozilla.com/D207687

UltraBlame original commit: 669dafe4d5a98758cd77f7b5150c45476734f373