Bug 1891185 - Collect different downgrade time on timer for HTTPS-First telemetry r=freddyb
It does not make sense to collect the "downgrade time" as we currently do it
when a upgrade was caused by a faster http request on a timer. We can not
compare the navigation start timestamp with the current time, as the replacing
http channel has started loading earlier than the current time. So leave out the
case where the downgrade happens because of a faster http request on the timer,
as we already know that the time overhead there is exactly 3s.

Differential Revision: https://phabricator.services.mozilla.com/D207338

UltraBlame original commit: 2deb5d0fae7d901200b6be63c53d156cc6e497ed