Bug 1817163 - Add a profiler feature for memory tracking to enable the memory hooks r=profiler-reviewers,perftest-reviewers,android-reviewers,kshampur,aabh,gl
This feature is added as "recommended", and added to every presets. This will
allow us to disable it if we really want to.

If the "native allocations" feature is enabled, we still force the installation
of the memory hooks even if we don't have this feature to not break the native
allocations feature.

Differential Revision: https://phabricator.services.mozilla.com/D206789

UltraBlame original commit: dc425d760d0918fee1c46e930cf3badd1f1f0fb9