Bug 1894249 - Remove duplicate test. r=xpcom-reviewers,mccr8
This was for testing a variant of ToInteger added in bug 53847
that used slightly different types for its parameters.

That variant was removed in bug 777292 (6e30103582c44ea1ee8c25747bd3bed6671c0011)
but the duplicate test was not.

Differential Revision: https://phabricator.services.mozilla.com/D209016

UltraBlame original commit: eb238c78c43d8917cbd8706709cc09de5ff26685