Bug 1893087 - Part 1: Relax assertion in RemoveManagee, r=ipc-reviewers,mccr8
With the new changes in bug 1724083, it is possible for an actor to be
destroyed during SetManagerAndRegister, which can lead to RemoveManagee being
called before the managee is even registered, leading to an assertion failure.
The easiest fix here is to relax this assertion, as we won't insert the entry
later in this failure case.

Differential Revision: https://phabricator.services.mozilla.com/D208936

UltraBlame original commit: 8ec16738af2257565dadd1db33976d8a1313e721