Bug 1878716: Fix some Hardware Concurrency Tests r=timhuang
It turns out that we had the override string wrong, but we
didn't notice because we happened to be only testing things
where it was supposed to be disabled anyway.

There was a test where this would have not been the case
(simplePBMFPPTest) - but this was only be tested when we were
_exempting_ it with ETP.

I added this scenario to a place where it would be tested
positively, and confirmed that it failed, then when I fixed
the override, started working again.

Differential Revision: https://phabricator.services.mozilla.com/D205913

UltraBlame original commit: 734befa016cd1002b076121ee744d0954dec5698