Bug 1894564: Add an explicit scrollbar-color for textarea ltr/rtl scrollbar position reftests. r=layout-reviewers,emilio
These reftests are just trying to verify that the scrollbar ends up on opposite
sides for a left-to-right vs. right-to-left textarea. They do that by means of
a not-equal reftest comparison, but that comparison fails if the system happens
to have transparent (or white) scrollbar tracks.

This patch simply forces an opaque and clearly-distinguishable scrollbar-color
for these textareas, to avoid this fragile/accidental dependency on system
characteristics.

(While I'm at it, I also linewrapped the 'style' attribute for readability and
to avoid making the lines too long in these testcases.)

Differential Revision: https://phabricator.services.mozilla.com/D209450

UltraBlame original commit: 4d431510961958e8efcb68445acbed4407b928f9