Bug 1895599 - Fix omnijar reading with the zip crate. r=gsvelto,supply-chain-reviewers
This crate is only used by the crash reporter and geckodriver, and
geckodriver is only used in testing (if I understand correctly).

In the future we will upstream changes to the zip crate which more
gracefully handles the case which they are trying to cover.

Differential Revision: https://phabricator.services.mozilla.com/D209752

UltraBlame original commit: deb3eaac74f476b6bc6c135f209cbcba2ac51f0d