Bug 1895434 - [devtools] Stop forcing the Performance panel to be displayed in the Browser Toolbox r=devtools-reviewers,nchevobbe
We used to force the preferences for the new performance panel to true for the Browser Toolbox to avoid showing the old performance panel, regardless of the preferences set by the user for their regular devtools.

However, now that the old performance panel is gone, it should be fine to let users decide if they want the panel or not via the devtools Settings panel.
Also the performance panel is on by default, so it shouldn't make it more complicated to profile on clean profiles.

Differential Revision: https://phabricator.services.mozilla.com/D210148

UltraBlame original commit: 91f506042b8a04e1259451ce1d25e4c9f0b241da