Bug 1895462 - Allow UniquePtrs to be used in WeakCache hash maps r=sfink
The problem here is that we copy the keys/values in entryNeedsSweep. I tried
several strategies to avoid this but in the end the simplest thing to do was
add a needsSweep method to the GC policy templates that calls
IsAboutToBeFinalized. I've implemented this only where necessary (and WeakCache
is only used internally nowadays).

Differential Revision: https://phabricator.services.mozilla.com/D209687

UltraBlame original commit: bb3c347722e6904d89d22631d3f965a3ee7a7e84