Bug 1880503 - Handle sync IPC timeout in UiCompositorControllerChild. r=aosmond
Extend the sync IPC timeout mechanism in CompositorManagerChild to
additionally cover UiCompositorControllerChild. As
UiCompositorControllerChild runs on the Android UI thread, we ensure
GPUProcessManager::KillProcess dispatches to the gecko main thread.

Along with the previous patch in this series this should provide us
with crash reports when the Android UI thread is hung waiting for the
GPU process to reply.

Differential Revision: https://phabricator.services.mozilla.com/D202167

UltraBlame original commit: 437063345a7ba4c220cdaa451796020ef5390cab