Bug 1893088 - Split coord space mapping for stacking context coord / external scroll offsets r=gfx-reviewers,nical
During display list and scene building, there are two coordinate
remapping steps that occur:

1) From stacking context coords -> reference frame relative coords
2) From pre-scrolled coords -> removed external scrolling offsets

These were previously handled in one place, however we want to
split these up so that we can apply snapping _after_ step 1 but
prior to step 2. This will allow us to have fractional external
scroll offsets that don't affect snapping. These will be snapped
later on during frame build after applying any (possibly fractional
APZ scroll offsets). This is a cheap operation during frame building
as we only need to snap and modify the transform matrices, not
individual primitives.

This patch should have no functional changes, it's prep work for
the changes referenced above. It does move all of step 1 to be
done during DL building in the content process, and all of step
2 to be done during scene building in the GPU process. In future,
if/when we resolve the issues we have with reliance on cross-iframe
knowledge for fractional snapping, we can move step 2 (including
snapping) in to the content process as well.

Further, as part of the DL bypass work, we will need to remap coord
spaces during DL building in a differeny way, which this simplifies.

Differential Revision: https://phabricator.services.mozilla.com/D208427

UltraBlame original commit: b5d01eb677b997d19730974df25440fac2f6449a