Bug 1844446 - [devtools] Don't use desugared selectors for selector highlighter. r=devtools-reviewers,bomsy.
Instead, we use the new CSSStyleRule.querySelectorAll method to retrieve the
elements to highlight.
We still compute a "selector", which is basically a concatenation of the
ancestor rules, so we can still mark the rule as highlighted when we display a
different StyleRule front for a rule that was previously highlighted.

Differential Revision: https://phabricator.services.mozilla.com/D208364

UltraBlame original commit: 6973a334460157977b8eca76c9fed772d266269c