Bug 1894605 - Fix verifyAboutFirefoxMenuItems UI test r=ohorvath
The UI test failed because the TCP CFR was displayed on the second and third try, after the test failed to verify the toolbar “Firefox Fenix | OSS Libraries” title on the first try.

This was happening because we were disabling certain settings in the test, and when the second and third try retried the test, the settings were actually set to the default values (in this case the TCP CFR was set to true. blocking the view)

Using the withDefaultSettingsOverrides function will ensure that the settings remain disabled even if the test will take use of the **retryTestRule**

Differential Revision: https://phabricator.services.mozilla.com/D210160

UltraBlame original commit: 4ba719fe24675d30851a12d300acd0e306dbc1e5