Bug 1875502 - Update UpdateServiceStub consumers r=nalexander,application-update-reviewers,firefox-desktop-core-reviewers
The earlier patches in this stack (a) made it possible to await on the stub to complete and (b) ensured that the update system initiates properly regardless of when and if the update service stub is invoked. This allows us to remove explict stub invocations in some cases and to have the await on the results in others.

Differential Revision: https://phabricator.services.mozilla.com/D209129

UltraBlame original commit: 924b7230a45ee4f3e2720e88cd4fc4a46fd625f3