Bug 1896538 - Enhance FFmpegEncoderModule's SupportsCodec check r=media-playback-reviewers,padenot
This patch enhances the `SupportsCodec` function by explicitly verifying
the supported encoder within the FFmpegEncoderModule. Previously, the
function checked whether the CodecType could be mapped to an AVCodecID,
but this alone didn't guarantee the existence of the corresponding
encoder. Now, the patch ensures the existence of the encoder by directly
checking for it within the FFmpegEncoderModule.

Differential Revision: https://phabricator.services.mozilla.com/D210252

UltraBlame original commit: 166e08b48e06eb80bfb2064ceae565464d5f7fdc