Bug 1894012 - Part 2: Store the AllocSite's kind explicitly r=jandem
Currently we work this out by other means but it's simpler to store it directly
using the top bits of the bytecode offset.

I figure it's OK to only use 30 bits to store the offset since we store it in
28 bits in RetAddrEntry, although I couldn't find it documented how many bits
are actually required.

Differential Revision: https://phabricator.services.mozilla.com/D210506

UltraBlame original commit: 47a8323caddab6310aed992cce5653b9c2005cc0