Bug 1898042 - Sort extra_labels in reportbrokensite for consistency. r=twisniewski,webcompat-reviewers
Not sorting this code means that tests are sensitive to the order of properties
in an object, or the order various test code sticks it into the array. It doesn't
look like the order actually matters, so sort extra_labels in the two places
that create these arrays.

Differential Revision: https://phabricator.services.mozilla.com/D211088

UltraBlame original commit: 4218099ea87336912f889951d5b270253c13f561