Bug 1896225 - Add the crashtest for bug 1896225 r=jjaschke
The crash itself was about when the selection crossed the
shadow boundary first, and then got moved to the same boundary later.

This bug was fixed in bug 1891783, so we just need to add the
crashtest.

Differential Revision: https://phabricator.services.mozilla.com/D211220

UltraBlame original commit: f775ae5b7dafe869866c64f26b8b74c8847c576d