Bug 1888429 - Part 1.5: Limit number of dispatched tasks for all task kinds r=jandem
This removes the check from checkTaskThreadLimit that always allows tasks to
start if |maxThreads == threadCount|. Previously there was a further check on
the number of dispatched tasks that limited this to the number of threads but
that was removed in the previous patch. I don't know why this check was ever
there.

Without this change we will dispatch helper thread tasks for all Ion
compilations which can result in a very large number of tasks being dispatched.
Combined with the fact that this patch stack means that we can't cancel tasks
after they've been dispatched but before they start running this meant that
some tests were timing out.

The patch also simplifies the rest of that method as I've always found it
confusing.

Differential Revision: https://phabricator.services.mozilla.com/D209918

UltraBlame original commit: 5cf3c43ba7e93bc08bec843afb221b0c6c7663e6