Bug 1896758 - Part 2. Add MediaDataEncoderProxy to wrap encoders that run on a particular thread. r=media-playback-reviewers,padenot
This patch adds a new class, MediaDataEncoderProxy, that is the
equivalent of MediaDataDecoderProxy for encoders. It allows using
encoders that must be accessed on a particular thread. It will be used
by a later patch in this series.

Differential Revision: https://phabricator.services.mozilla.com/D211651

UltraBlame original commit: ee8a31fcc5d7434a8e5de7c59de902f3e676a4ad