Bug 1899210 - Fix crash in ConfigureMessage::ToString r=media-playback-reviewers,padenot
This patch adds a `nsCString` member to store the codec name in the
`ConfigureMessage` class for use in its `ToString` method. Previously,
`ToString` generated a codec string from the stored config. However,
when `TakeConfig` is called, the config is moved out, leaving a null
config in the class. If `ToString` is called afterward, it causes a
crash because the code attempts to dereference the null config to obtain
the codec name.

Differential Revision: https://phabricator.services.mozilla.com/D211801

UltraBlame original commit: 747eed3ff86eae71ec594e4288559e54a5ffaeef