Bug 1898103 - Optimize config sharing to avoid copies r=media-playback-reviewers,padenot
This patch optimizes the handling of configuration information in the
`DecoderTemplate` by replacing the `UniquePtr` that stores the current
configuration with a `RefPtr`. This change allows the configuration to
be shared within the tasks running the output callbacks, rather than
being copied. As a result, memory usage and performance are improved.

Differential Revision: https://phabricator.services.mozilla.com/D211272

UltraBlame original commit: 9c18eeef30288e38bc857841bedca1bca75505fb