Bug 1898473 - Don't access shared permanent atoms zone when marking symbols r=sfink
This is complaining that a worker runtime is accessing the shared permanent
atoms zone, which is owned by the main runtime. This is not a problem because we
only check the zone GC state here and the zone will never be collected while the
worker is running.

The patch adds a test to avoid the assertion failure.

Differential Revision: https://phabricator.services.mozilla.com/D211822

UltraBlame original commit: 1985b0103a3a70a18a0655ef1f8fecbfb1c3ff64