Bug 1899699 - [devtools] Display computed value for properties used in property value. r=devtools-reviewers,bomsy.
In the computed panel, by default, we're only showing the properties
that are set in the matching rules for the selected element.
With registered custom properties, we can have cases where the
property is not set, in which case, its initial value will be used.
It will probably be helpful to show the computed value of any
custom property that is used on the selected node.

Depends on D212092

Differential Revision: https://phabricator.services.mozilla.com/D212099

UltraBlame original commit: eadaafd5fd0315f7a08f407312a3f719d1ff7ccd