Bug 1841370 - Add `mach repackage snap` and `snap-install` for testing development builds under Snap. r=firefox-build-system-reviewers,sergesanspaille
This adds the command `mach repackage snap`, which packages the current
build as a Snap (with the app name firefox_devel, to not interfere with
installs of the normal firefox package).  It can also accept a tar file
as input, for example to repack a build from Taskcluster.

It also adds `mach repackage snap-install`, which installs the package
and connects its plugs with `snap connect` (normally done by default,
but not for an unsigned package like this).

Building a Snap package requires the `snapcraft` command, which is
available as a Snap (and maybe from other package managers).

Depends on D182615

Differential Revision: https://phabricator.services.mozilla.com/D182616

UltraBlame original commit: 051187f2c76e1affe3df0017b13b3928c3c5f8f0