Bug 1892734 - Add tests for checking the priority set in our HTTP requests. r=valentin
In bug 1864392, the urgency parameter set in our HTTP requests [1] is
adjusted according to the value of the internal priority. This patch
adds basic tests to verify some adjustments, and indirectly the internal
priority, for scripts, images, styles, fonts and global fetches using
different values of fetchpriority.

This is done by introducing a Python handler [2] that serves different
kind of content depending on the urgency parameter it receives and a
WPT test that extracts the urgency from the served content.

[1] https://www.rfc-editor.org/rfc/rfc9218.html#name-urgency
[2] https://web-platform-tests.org/writing-tests/python-handlers/index.html

Differential Revision: https://phabricator.services.mozilla.com/D211988

UltraBlame original commit: b1240a12bbe15edf41fc2cf2d09cec387988dd44