Bug 1839637, part 2 - Stop warning in HandleEvent when owner is not an inner window. r=nika
After the fix from the previous patch, this warning will be the third most common
warning, with over 41,000 instances, on Linux 64. I'm not sure if the owner global is
null, or if it is something like a BackstagePass.

Differential Revision: https://phabricator.services.mozilla.com/D212736

UltraBlame original commit: 8a918f98bbe3bf161247b39d799754bba7e0582f