Bug 1892192 - [devtools] Display starting-style rules in Rules view. r=devtools-reviewers,bomsy.
This takes care of displaying `starting-style` rules in the Rules view.
A test is added to cover this, as well as a test case to make sure those
starting style rules are not impacting the computed panel.

Depends on D210363

Differential Revision: https://phabricator.services.mozilla.com/D209329

UltraBlame original commit: b045155b8af1bd6225296d67bf6c58c9710a0918