Bug 1867605 - [devtools] Record actual frame durations in JS tracer profiler data. r=canaltinova,devtools-reviewers,nchevobbe
The key here is to record a sample at the end of frame execution.
This helps the profiler frontend detect the end of the currently execution frame.
When we enter a frame (start executing a function), we record a first sample
about that new stack matching that newly executing frame's location (=function location).
Then, when we exit the frame (function ends running), we record another sample
which will be the parent stack. i.e. the stack that was running before that function call.
This stack may be "null". In which case, nothing is currently executing anymore.

Differential Revision: https://phabricator.services.mozilla.com/D195128

UltraBlame original commit: c6227883e96f434c4f3c9952294e7b62c60e4212