Bug 1893526 – Change TabCounterToolbarButton visibility based on that of navbar r=android-reviewers,petru
The intent here is to show tabs icon in the toolbar in landscape mode and for tablets. Previously, we completely disabled tabs button for the redesigned version (with the navbar), but the condition is more nuanced – even the redesigned version shows tabs tray in case it's in landscape mode or the app is running on a tablet. Basically, the rule is – if there is not navigation bar (that has it's own tabs counter button), toolbar is displaying tabs counter.

Differential Revision: https://phabricator.services.mozilla.com/D211590

UltraBlame original commit: fce71e4ecf9be4d367dc20c128059050a9e6328c