Bug 1884378 - Introduce SandboxProfiler r=padenot
Provide a way to forward sandbox decisions and syscalls interceptions to
the profiler, with child-side stack in the reports.

Introduces a new SandboxProfilerEmitter thread that will be able to deal
with what cannot be done in the context of a signal handler.

Child-side syscalls interceptions with stacks requires a bit of care
because of the signal context it is running under. Taking a stack is
safe in that context but dealing with the profiler is not, thus relying
on the MPSCQueue data structure:
 - when in SIGSYS because a syscall has been intercepted, take a stack
 - forward the syscall informations plus the stack within the MPSCQueue
 - SandoxProfilerEmitter can then read from the queue and perform the
   heavy work with the profiler

Differential Revision: https://phabricator.services.mozilla.com/D207736

UltraBlame original commit: 9ca5a7a92e64c3520c19b0f4088a1347218ee2eb