Bug 1902157 - Throw NotSupportedError for unsupported VideoFrame formats r=media-playback-reviewers,padenot
This patch updates the exception type thrown for unsupported formats.
Previously, a `TypeError` was thrown when the specified `VideoFrame`
format was unsupported. Now, the patch changes this to throw a
`NotSupportedError`, which more accurately reflects the nature of the
error.

Differential Revision: https://phabricator.services.mozilla.com/D213730

UltraBlame original commit: 916a5b8267fccc4b0de2f6ad2426aefdb0fcc1f1