Bug 1824053 - [1/4] Remove Win32 implementation of nsIWidget::SetBackgroundColor()  r=win-reviewers,handyman
This never worked according to its description: it merely set the
_window-class_ background-color. We neither have nor need any mechanism
to give individual windows their own per-widget background color under
Win32.

Instead, we just set the background color to be dark grey, which is
probably good enough for the few purposes for which the background brush
is actually used. (Alas, this does not _quite_ include the initial white
flash.)

The comment in this bug about this being important for Windows dates
from bug 574638, and is no longer relevant -- at least, not in the
particular way it used to be.

Differential Revision: https://phabricator.services.mozilla.com/D212761

UltraBlame original commit: 49613184d35923575ab6cd3a9ffb544678730207