Bug 1896718 - Implement the prefFlips feature r=chumphreys
This patch implements the prefFlips feature that was added as a placeholder in
bug 1896720. The schema has been updated to fix a bug, as the accepted values
for the `branch` key were incorrect. Additionally, the schema now forbids pref
flips that would clear the default branch, as that is something we cannot do.

When a setPref experiment is currently enrolled and a prefFlips experiment
would enroll that touches the same prefs, the setPref experiment is unenrolled
and vice versa (i.e., the old experiment will always unenroll). This is an edge
case, as we do not expect this to happen in practice (since if a pref is
covered by a feature with setPref, the prefFlips feature probably shouldn't be
used to change it). Unenrollment telemetry has been updated to add a new
"prefFlips-conflict" reason for this case, which also includes the slug of the
experiment that caused unenrollment (as `conflictingSlug` / `conflicting_slug`
in legacy and Glean, respectively).

Differential Revision: https://phabricator.services.mozilla.com/D213506

UltraBlame original commit: 2a746899fe0e56bade0179f0f5fad49f3f2b9b89