Bug 1891182 - part 01 - rename ModuleEnvironment to ModuleMetadata.  r=rhunt.
This patch is huge, but it does only one thing: renames ModuleEnvironment to
ModuleMetadata, and consistently renames variables etc of that type to
`moduleMeta` / `moduleMeta_` / `moduleMeta()`.  No functional change.

This is intended to be a preliminary to splitting out a new type,
CodeMetadata, and associated `codeMeta` names, in a later patch in this
series.

Differential Revision: https://phabricator.services.mozilla.com/D211157

UltraBlame original commit: 14d7c72951a190267dd0f101db97b57f6a247682