Bug 1888756, part 3 - Text Fragments: Add a flag in the `DocShell` to keep the information if a load was same-document. r=farre,dom-core,edgar
This commit is a prerequisite for part 5 of this patch set.

The idea is to keep the information whether the last
load done by the doc shell was a same-document navigation.
The value is reset for every navigation, therefore it
should always be correct.

There is no changed behavior in this commit.

Differential Revision: https://phabricator.services.mozilla.com/D212818

UltraBlame original commit: 400a8ff57ae4c7af776cace3dab9f2f956c21a89