Bug 1901851 - Part 1: Add BackgroundParent::GetBackgroundThread, r=dom-worker-reviewers,ipc-reviewers,asuth,mccr8
Previously the identity of the BackgroundParent thread was hidden to encourage
all communication with that thread, including communication within the parent
process, to go over PBackground. This has not born out as effective in
practice, however.

By exposing the thread's identity through a getter available on the background
or main thread, we can simplify the process of dispatching a runnable to the
background thread. This will be useful later in this bug as some actors are
migrated off of PBackground, but still need to run on that thread.

Differential Revision: https://phabricator.services.mozilla.com/D213311

UltraBlame original commit: b29282956a04fb6d0999f36bb6e8810216755f48