Bug 1860915, part 1 - Implement part of Document Policy for force-load-at-top. r=necko-reviewers,smaug,kershaw
This adds support for a piece of the Document Policy header parsing we need for
force-load-at-top, and sets a bit on Document if text fragments are enabled,
but this patch does not actually check the bit. That is done in the next
patch.

Differential Revision: https://phabricator.services.mozilla.com/D191798

UltraBlame original commit: c93c2e13371c1aa1271ce11f60e0079dd26a2898