Bug 1891182 - part 15 - move ModuleElemSegmentVector from CodeMetadata to ModuleMetadata.  r=rhunt.
This patch moves ModuleElemSegmentVector from CodeMetadata to ModuleMetadata.
A vector of RefTypes is added to CodeMetadata, and this is a copy of the
.elemType fields in the ModuleElemSegmentVector.

When a wasm::Module is finally created, the ModuleElemSegmentVector is
std::move'd into the Module -- this is functionality that already exists.
Prior to this patch, the vector was moved from the CodeMetadata; now it is
moved from the ModuleMetadata.  It is intended that a future patch will remove
the duplicated field in Module, hence the move, and so the
ModuleElemSegmentVector will be used directly from its new home in
ModuleMetadata.

Differential Revision: https://phabricator.services.mozilla.com/D214130

UltraBlame original commit: e624ca90c9bfd944f8c2d3bf3b629c4a52243d05