Bug 1903922 - Add an upward-pointing triangle glyph to -moz-bullet-font. r=layout-reviewers,emilio
Updated using the current version (6.6.0) of BirdFontNonCommercial, https://birdfont.org.

The only "real" modification from Mats' original bullet-font is the addition of the glyph for U+25b4;
other changes are an artifact of editing the source with the updated tool. In particular, I notice
that it has saved floating-point values with lower (but still more-than-adequate) precision.
Given that we end up generating a 1024-upem TrueType font, five decimals of precision in
the path data (the tool's current default) is plenty.

Differential Revision: https://phabricator.services.mozilla.com/D214554

UltraBlame original commit: 147da497b73550e48142477d4e4550f4b7923098