Bug 1894228 - Change the aOutFirstInput parameter of InputQueue::FindBlockForId() from a pointer to an iterator. r=dlrobertson
This allows the caller to iterate the input queue to look at events
beyond the first event in the block, if desired.

To avoid a circular dependency between InputBlockState.h and InputQueue.h,
a class InputQueueIterator is introduced which hides the concrete type
of the iterator.

Differential Revision: https://phabricator.services.mozilla.com/D212977

UltraBlame original commit: e0e8173848faeb33787549307a5048cd843d7b8a