Bug 1904096 - revert clang-formatting from GeckoTraceEvent.h r=ng,profiler-reviewers,aabh
This file is marked in .clang-format-ignore because it
is copied from libwebrtc.  Formatting changes here make
it difficult to keep it in step with upstream changes.

The file now contains '// clang-format off' and
'// clang-format on' to help avoid this issue in the
future.

Differential Revision: https://phabricator.services.mozilla.com/D214620

UltraBlame original commit: 562736492281b18ae43e738991ed14637cad46a6