Bug 1898745 - P1. Store DOM content reference id in FieldDetail structure r=NeilDeakin,credential-management-reviewers
In order to run the field detection heuristics in the parent process,
the parent requires a unique identifier for DOM elements to facilitate
communication with the child.

Therefore, we utilize the id from ContentDOMReference as the unique identifier.
This identifier is stored in the FieldDetail structure, which the formautofill
module uses to maintain all information related to an identified field.

Differential Revision: https://phabricator.services.mozilla.com/D211550

UltraBlame original commit: 7b7441408b33d431b53749f981b240ae469fbc0d