Bug 1898745 - P3. Record form interactio telemetry in the parent process r=credential-management-reviewers,NeilDeakin
We have moved most of the logic from the child to the parent. This patch updates
the locations where telemetry is now recorded in the parent. Additionally,
it updates the API of AutofillTelemetry to comply with the new architecture.

Differential Revision: https://phabricator.services.mozilla.com/D211552

UltraBlame original commit: 576634ec09aacc12925cfd08051d5d5f046e56ba